fix query cache in embedded
this fixes main.partition_cache and main.cache_innodb in --embed

followup for 430d60d1fc0 MDEV-24487

+++ libmysqld/embedded_priv.h
@@ -19,8 +19,8 @@
 
 C_MODE_START
 void lib_connection_phase(NET *net, int phase);
-void init_embedded_mysql(MYSQL *mysql, int client_flag);
-void *create_embedded_thd(int client_flag);
+void init_embedded_mysql(MYSQL *mysql, ulong client_flag);
+void *create_embedded_thd(ulong client_flag);
 int check_embedded_connection(MYSQL *mysql, const char *db);
 void free_old_query(MYSQL *mysql);
 extern MYSQL_METHODS embedded_methods;
