+++ libmysqld/lib_sql.cc
@@ -660,7 +660,7 @@ void end_embedded_server()
 }
 
 
-void init_embedded_mysql(MYSQL *mysql, int client_flag)
+void init_embedded_mysql(MYSQL *mysql, ulong client_flag)
 {
   THD *thd = (THD *)mysql->thd;
   thd->mysql= mysql;
@@ -680,7 +680,7 @@ void init_embedded_mysql(MYSQL *mysql, int client_flag)
   create_new_thread(), and prepare_new_connection_state().  This should
   be refactored to avoid code duplication.
 */
-void *create_embedded_thd(int client_flag)
+void *create_embedded_thd(ulong client_flag)
 {
   THD * thd= new THD(next_thread_id());
